Add vercel for deployment. Rename server to api folder. Fix token server.
