+++ .gitignore
@@ -33,3 +33,5 @@ junit.xml
 serviceAccountKey.json
 
 public/virtualbackground/
+
+.vercel
