+++ api/utils/auth-handler.ts
@@ -1,0 +1,7 @@
+import { RequestHandler } from 'express';
+
+const authMiddleware: RequestHandler = async (req, res, next) => {
+  next();
+};
+
+export default module.exports = authMiddleware;
