+++ api/utils/bootstrap-globals.ts
