+++ api/utils/firebaseAuthMiddleware.ts
