--- server/auth-handler.ts
@@ -1,1 +1,0 @@
-export default module.exports = () => {};
