+++ tailwind.config.js
@@ -16,8 +16,5 @@ module.exports = {
         scale: ['group-focus']
       },
     },
-    plugins: [
-      require('@tailwindcss/aspect-ratio'),
-    ],
   }
 }
